Move to proper maven structure (so mvn clean package works)
